Rename workflow history files
